backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Google AI & ADK
google-adk
google-genai
google-generativeai

# Database
aiosqlite
sqlalchemy

# Utilities & Tools
python-dotenv
requests
beautifulsoup4
PyPDF2
pydantic-settings